Moved classes to it's own files
