+++ src/discogs/app/draft/types.ts
@@ -1,0 +1,5 @@
+export type FormElement =
+  | HTMLInputElement
+  | HTMLSelectElement
+  | HTMLTextAreaElement
+  | null;
