+++ src/discogs/app/draft/variation.ts
@@ -1,0 +1,11 @@
+export class Variation {
+  value: string;
+
+  constructor(value: string) {
+    this.value = value;
+  }
+
+  toString(): string {
+    return this.value;
+  }
+}
